022.04 Changes to motor power function

+++ TinyG2/stepper.h
@@ -270,7 +270,7 @@ typedef struct stRunMotor { 		// one per controlled motor
 	int32_t phase_accumulator;		// DDA phase angle accumulator for axis
 	uint8_t power_state;			// state machine for managing motor power
 	uint32_t power_systick;			// sys_tick for next state transition
-	uint32_t power_level;			// power level for this segment (FUTURE)
+	float power_level;				// power level for this segment
 	uint8_t step_count_diagnostic;	// step count diagnostic
 } stRunMotor_t;
